project/requirements.txt
pip==21.1.1
setuptools==56.1.0
spotipy~=2.18.0
pandas~=1.2.4
numpy~=1.20.2
databases
pydantic~=1.8.1
fastapi==0.64.0
starlette~=0.13.6
requests~=2.25.1
h11~=0.12.0
pytz~=2021.1
python-dateutil~=2.8.1
SQLAlchemy
asyncpg~=0.22.0
certifi~=2020.12.5
chardet~=4.0.0
six~=1.15.0
urllib3~=1.26.4
click~=7.1.2
uvicorn~=0.13.4
idna~=2.10
python-multipart~=0.0.5
tortoise-orm
pytest
gunicorn